Update Booze Clues Code
Created the page that shows small pictures of what each bars look like as well as tips when you mouse hover over the picture that gives a brief description of each one. Next is to give the user the option to click the bar they are interested in and see where in town it is located as well as events/music/vibes/ and what they serve at that given bar. The code doesn't give much but the interface itself shows what is going on.