Testing out streamwriter
I was able to link the text document with the actual program. When the user hits the "SIGNUP" button, it'll save the "Hello" into the text file. The goal for Friday is to be able to save data that the user enters in the text boxes that include their first and last name, username and password. The "Hello" was just text code to make sure the path was working properly.