Update Booze Clues Code
What I did was I made some of the descriptions a tad shorter and moved some of the details into bullet points of some of the things those bars consisted of. I went through each one and added their hours for each day so the customer knows when they open and close. I spent some time attempting to design the form that includes all this information.  Changes are still being made to the form that is pleasing to the eye, yet easy to read.