Update Booze Clues Code
When the user clicks on a picture of the bar they are interested in going to, it will bring them to page that have a description of it, the hours, what type of things/vibess are listed and a link that will send them to either their website or Facebook page. Buttons were also added to direct the user back to the other bars, the restaurants page, or to close the app.