Links to HTML Pages
Instead of each picture block of the bars linking to another form that displays the details, we decided to create HTML Pages that those pictures will now link to. This will be more user friendly as well as giving us more room to add in more details and information than on that small form. Using HTML will refresh my memory and also teach the group how to use it.