Update HTML Bar Specials
Bits and pieces of the <Style> were tweaked and changed so I just copy and pasted the entire style tag instead of going through the Git code and changing piece by piece. The container was changed so the page is more narrow as well as simplified. The header and Footer are no longer a block of color, just a solid grey border below the header and a solid grey border above the footer. I removed the Spotlight picture and plan on adding a low contrasted background image to the footer. This change will occur to all the bar description pages, I just only changed Green Rooms in Git to show the changed.